refactor(task): 将任务 ID 的数据类型改为 int64
- 在 flags.go 中将 taskIDFlag 的类型从 IntFlag 改为 Int64Flag
- 在 db.go 中将 GetTaskNameByID 和 GetTaskByID 函数的参数类型从 int 改为 int64
- 这些修改确保了任务 ID 的数据类型在整个系统中保持一致，使用更合适的 int64 类型

+++ cmd/subcmd/run/flags.go
@@ -11,7 +11,7 @@ var (
 	runCmd *cmd.Cmd // run命令
 
 	// 任务选择参数
-	taskIDFlag   *qflag.IntFlag   // -id: 指定任务ID
+	taskIDFlag   *qflag.Int64Flag // -id: 指定任务ID
 	taskIDsFlag  *qflag.SliceFlag // -ids: 指定多个任务ID
 	allTasksFlag *qflag.BoolFlag  // -all: 运行所有任务
 )
@@ -23,7 +23,7 @@ func InitRunCmd() *cmd.Cmd {
 	runCmd.SetUseChinese(true)
 
 	// 任务选择参数（互斥）
-	taskIDFlag = runCmd.Int("", "id", 0, "指定要运行的任务ID")
+	taskIDFlag = runCmd.Int64("", "id", 0, "指定要运行的任务ID")
 	taskIDsFlag = runCmd.Slice("", "ids", []string{}, "指定多个任务ID进行批量运行")
 	allTasksFlag = runCmd.Bool("", "all", false, "运行所有任务")
